Feature/score feats

+++ digital-repository-back/src/main/java/co/unicauca/digital/repository/back/domain/repository/User/ICredentialRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
  * database.
  */
 @Repository
-public interface ICredentialRepository extends JpaRepository<Credential, Integer> {
+public interface ICredentialRepository extends JpaRepository<Credential, Long> {
 }
